Impl UserService & Impl ProductService

+++ src/main/java/entity/Boots.java
@@ -6,11 +6,11 @@ public class Boots extends Product {
     private boolean isNaturalSkin;
 
 
-    public Boots(int id, String productName, Float price, Float weight, String color, Integer productCount, Integer size,
+    public Boots(Long id, String productName, Float price, Float weight, String color, Integer productCount, Integer size,
                  boolean isNaturalSkin) {
         super(id, productName, price, weight, color, productCount);
-        this.size=size;
-        this.isNaturalSkin=isNaturalSkin;
+        this.size = size;
+        this.isNaturalSkin = isNaturalSkin;
     }
 
     public Integer getSize() {
@@ -23,7 +23,7 @@ public class Boots extends Product {
 
     @Override
     public String toString() {
-        return super.toString()+" Boots{" +
+        return super.toString() + " Boots{" +
                 "size=" + size +
                 ", isNaturalSkin=" + isNaturalSkin +
                 '}';
