+++ src/main/java/entity/Cloth.java
@@ -1,18 +1,16 @@
 package entity;
 
-public class Cloth extends Product{
+public class Cloth extends Product {
 
     private String size;
     private String material;
 
 
-
-
-    public Cloth(Integer id, String productName, Float price, Float weight, String color, Integer productCount, String size,
+    public Cloth(Long id, String productName, Float price, Float weight, String color, Integer productCount, String size,
                  String material) {
         super(id, productName, price, weight, color, productCount);
-        this.size=size;
-        this.material=material;
+        this.size = size;
+        this.material = material;
     }
 
     public String getSize() {
@@ -25,7 +23,7 @@ public class Cloth extends Product{
 
     @Override
     public String toString() {
-        return super.toString()+ " Cloth{" +
+        return super.toString() + " Cloth{" +
                 "size='" + size + '\'' +
                 ", material='" + material + '\'' +
                 '}';
